+++ android/src/main/kotlin/com/yandex/mobile/ads/flutter/appopenad/AppOpenAdLoaderCommandHandlerProvider.kt
@@ -53,7 +53,10 @@ internal class AppOpenAdLoaderCommandHandlerProvider(
             override fun onAdLoaded(appOpenAd: AppOpenAd) {
                 val listener = AppOpenAdEventListener()
                 appOpenAd.setAdEventListener(listener)
-                if (!hasListener) return
+                if (!hasListener) {
+                    appOpenAd.setAdEventListener(null)
+                    return
+                }
                 val id = adCreator.createFullScreenAd(APP_OPEN_AD, listener) { onDestroy ->
                     AppOpenAdCommandHandlerProvider(
                         ObjectHolder(appOpenAd),