+++ android/src/main/kotlin/com/yandex/mobile/ads/flutter/rewarded/RewardedAdLoaderCommandHandlerProvider.kt
@@ -54,7 +54,10 @@ internal class RewardedAdLoaderCommandHandlerProvider(
             override fun onAdLoaded(rewarded: RewardedAd) {
                 val listener = RewardedAdEventListener()
                 rewarded.setAdEventListener(listener)
-                if (!hasListener) return
+                if (!hasListener) {
+                    rewarded.setAdEventListener(null)
+                    return
+                }
                 val id = adCreator.createFullScreenAd(REWARDED_AD, listener) {
                     RewardedAdCommandHandlerProvider(
                         ObjectHolder(rewarded),